fix: move test to textWysiwyg.test.tsx
